// angular/angular-ngrx/src/app/core/models/user.ts
export class User {
  id: number;
  username: string;
  first_name: string;
  last_name: any;
  about: string;
  email: string;
  avatar_url: string;
  phone_number: number;
  password: string;
  country_code: string;
  country: any;
  city: any;
  street: string;
  state: string;
  postal_code: string;
}